0.15
- Added "no definition found" exception for !wiki,  changed the format
- Added !help all
- Changed !manage sleep/wake to ayybot.sleep/wake
- Grammar fixes
- Added state_sleep to settings.ini
- Moved filtering words to config.py
- Changed !listmembers to !members
- Excluded the bot from spam check

+++ ayybot.py
@@ -1,27 +1,27 @@
 import discord
 from random import randint
-from config import mail,password,things,eightball,helpmsg1,creditsmsg,jokemsg,memelist,quotes
+from config import mail,password,useful,meme,eightball,helpmsg1,creditsmsg,jokemsg,memelist,quotes,filterwords
 from datetime import timedelta, datetime
-import time
-import configparser
+import time,configparser,os,wordfilter,wikipedia,requests
 from giphypop import translate
-import wordfilter
-import os
+from bs4 import BeautifulSoup
 
 __title__ = 'AyyBot'
 __author__ = 'DefaltSimon with discord.py api'
-__version__ = '0.12'
+__version__ = '0.15'
 
 # import logging
 # logging.basicConfig(level=logging.DEBUG)
+
 client = discord.Client()
 clientchannel = discord.Channel()
-botvs = "0.12"
+botvs = "0.15"
+
 parser = configparser.ConfigParser()
 parser.read("settings.ini")
-
-with open("status.txt","w+") as file:
-    file.write("state_sleep=0")
+parser.set("Settings","state_sleep","0")
+with open('settings.ini', 'w') as configfile:
+    parser.write(configfile)
 
 def runme():
     client.run()
@@ -32,9 +32,6 @@ def loginme():
 def deletemsg(message):
     client.delete_message(message)
 
-def printout(message, disauthor):
-    print('{msg} was executed by {usr}'.format(msg=str(message.content), usr=disauthor))
-
 def gettime(time_elapsed):
     sec = timedelta(seconds=time_elapsed)
     d = datetime(1, 1, 1) + sec
@@ -42,7 +39,7 @@ def gettime(time_elapsed):
     return this
 
 def logdis(message):
-    if parser.getboolean("SettingsOne", "WriteLogs") == 1:
+    if parser.getboolean("Settings", "WriteLogs") == 1:
         with open('log.txt',"a") as file2:
             one = str(time.strftime("%Y-%m-%d %H:%M:%S", ))
             str2 = one + str(" : {msg} by {usr}\n".format(msg=message.content,usr=message.author))
@@ -50,9 +47,7 @@ def logdis(message):
 
 def checkwords(message):
     cmsg = str(message.content).lower()
-    with open('filterwords.txt') as f:
-        mylist = [line.rstrip('\n') for line in f]
-    wordfilter.add_words(mylist)
+    wordfilter.add_words(filterwords)
     if wordfilter.blacklisted(cmsg) and message.author != "AyyBot":
         client.send_message(message.channel, "@{usr}, watch it!".format(usr=message.author))
 
@@ -67,7 +62,7 @@ def checkspam(message):
             continue
         if c == spamword:
             another += 1
-    if another > 7:
+    if another > 7 and message.author != "AyyBot":
         client.delete_message(message)
         client.send_message(message.channel,
                             "@{usr} Spam is not allowed. **Deal with it** ( ͡° ͜ʖ ͡°)".format(usr=message.author))
@@ -76,69 +71,66 @@ def checkspam(message):
 @client.event
 def on_message(message):
     try:
-        state_sleep = (open("status.txt","r").read()).split()[0]
+        state_sleep = int(parser.getboolean("Settings","state_sleep"))
         messagestr = str(message.content).lower()
         disauthor = message.author
         # check for !manage before (possibly) quiting out of def
-        if messagestr.startswith("!manage"):
-            msg6 = str(message.content[8:])
-            print(msg6)
+        if messagestr.startswith("ayybot."):
+            msg6 = str(message.content[7:])
             if msg6 == "sleep":
                 if state_sleep == 1:
                     client.send_message(message.channel,"No need m8, I was already sleeping.")
-                open("status.txt","w+").write("state_sleep=1")
+                    return
+                parser.set("Settings","state_sleep","1")
+                with open('settings.ini', 'w') as configfile:
+                    parser.write(configfile)
                 client.send_message(message.channel,"Now sleeping...")
                 print("Saved sleeping")
             elif msg6 == "wake":
-                open("status.txt","w+").write("state_sleep=0")
+                parser.set("Settings","state_sleep","0")
+                with open('settings.ini', 'w') as configfile:
+                    parser.write(configfile)
                 client.send_message(message.channel,"Returned from winter sleep.")
                 print("Returned to normal")
-        if state_sleep == "state_sleep=1":
+        if state_sleep == 1:
             print("Sleeping...")
             return
         # Spam and swearing check / with settings.txt check in config.checksettings(parm=1 or 2)
-        if parser.getboolean("SettingsOne", "FilterWords") == 1:
+        if parser.getboolean("Settings", "filterwords") == 1:
             checkwords(message)
-        if parser.getboolean("SettingsOne", "FilterSpam") == 1:
+        if parser.getboolean("Settings", "filterspam") == 1:
             checkspam(message)
         # commands imported from config.py
-        for onething in things.keys():
+        for onething in meme.keys():
             if message.content.lower().startswith(onething):
                 deletemsg(message)
-                second = things.get(str(message.content.lower()))
-                print(str("{msg} by {usr}".format(msg=message.content,usr=message.author)))
+                second = meme.get(str(message.content.lower()))
+                print(str("{} by {}".format(message.content,message.author)))
+                logdis(message)
+                client.send_message(message.channel, second.format(usr=disauthor.id))
+                return
+        for onething2 in useful.keys():
+            if message.content.lower().startswith(onething2):
+                second = useful.get(str(message.content.lower()))
+                print(str("{} by {}".format(message.content,message.author)))
                 logdis(message)
                 client.send_message(message.channel, second.format(usr=disauthor.id))
                 return
         if messagestr.startswith("!help"):
-            deletemsg(message)
             if messagestr.startswith("!help useful") or messagestr == "!help":
                 client.send_message(message.channel, helpmsg1)
             elif messagestr.startswith("!help fun"):
                 client.send_message(message.channel, jokemsg)
             elif messagestr.startswith("!help meme") or messagestr.startswith("!help memes"):
                 client.send_message(message.channel, memelist)
-            print(str("{msg} by {usr}".format(msg=message.content,usr=message.author)))
+            elif messagestr.startswith("!help all"):
+                client.send_message(message.channel, helpmsg1 + jokemsg + memelist)
+            print(str("{} by {}".format(message.content,message.author)))
             logdis(message)
-        # Restart
-        elif messagestr.startswith("ayybot.reboot"):
-            role1 = discord.utils.find(lambda role: role.name == 'serveradmins', message.channel.server.roles)
-            role2 = discord.utils.find(lambda role: role.name == 'mods', message.channel.server.roles)
-            role3 = discord.utils.find(lambda role: role.name == 'developers', message.channel.server.roles)
-            if message.author.id == message.channel.server.owner.id or role1 or role2 or role3:
-                deletemsg(message)
-                client.send_message(message.channel, "<@" + disauthor.id + "> Restarting...")
-                print(str("{msg} by {usr}, restarting".format(msg=message.content,usr=message.author)))
-                logdis(message)
-                os.system("python launchbot.py")
-            else:
-                logdis(message)
-                print(str("{msg} by {usr}, but incorrect permissions".format(msg=message.content,usr=message.author)))
         # Shut down
         elif messagestr.startswith("ayybot.kill"):
             role1 = discord.utils.find(lambda role: role.name == 'serveradmins', message.channel.server.roles)
             if message.author.id == message.channel.server.owner.id or role1:
-                deletemsg(message)
                 client.send_message(message.channel, "**DED**")
                 print(str("{msg} by {usr}, quitting".format(msg=message.content,usr=message.author)))
                 logdis(message)
@@ -151,29 +143,33 @@ def on_message(message):
             deletemsg(message)
             msg5 = message.content.lower()[6:]
             client.send_message(message.channel, "<@" + disauthor.id + "> rolled :" + str(randint(0, int(msg5))))
-            print(str("{msg} by {usr}".format(msg=message.content,usr=message.author)))
+            print(str("{} by {}".format(message.content,message.author)))
             logdis(message)
         # Dice - 1 - 6
         elif messagestr.startswith("!dice"):
             deletemsg(message)
             client.send_message(message.channel, "<@" + disauthor.id + "> You got: " + str(randint(1,6)) + "!")
-            print(str("{msg} by {usr}".format(msg=message.content,usr=message.author)))
+            print(str("{} by {}".format(message.content,message.author)))
             logdis(message)
+        # Game
+        elif messagestr.startswith("!game"):
+            deletemsg(message)
+            msgcut = messagestr[5:]
+            print(msgcut)
+            client.send_message(message.channel,"@everyone Does anyone want to play {}? ({})".format(str(msgcut),disauthor))
         # Credits
         elif messagestr.startswith("!credits"):
             deletemsg(message)
-            client.send_message(message.channel, creditsmsg)
-            print(str("{msg} by {usr}".format(msg=message.content,usr=message.author)))
+            client.send_message(message.channel, creditsmsg.format(bot=str(botvs)))
+            print(str("{} by {}".format(message.content,message.author)))
             logdis(message)
         # Cats are cute
         elif messagestr.startswith("!cats"):
-            deletemsg(message)
             client.send_file(message.channel, "images/cattypo.gif")
-            print(str("{msg} by {usr}".format(msg=message.content,usr=message.author)))
+            print(str("{} by {}".format(message.content,message.author)))
             logdis(message)
         # Lists all members in current server
-        elif messagestr.startswith("!listmembers"):
-            deletemsg(message)
+        elif messagestr.startswith("!members"):
             client.send_message(message.channel, "<@" + disauthor.id + "> **Current members:**")
             count = 0
             members = ''
@@ -186,23 +182,36 @@ def on_message(message):
             final = "**Total : {count1} members.**".format(count1=count)
             client.send_message(message.channel, members)
             client.send_message(message.channel, final)
-            print(str("{msg} by {usr}".format(msg=message.content,usr=message.author)))
+            print(str("{} by {}".format(message.content,message.author)))
             logdis(message)
+        # Online users
+#        elif messagestr.startswith("!online"):
+#            count = 0
+#            memberson = ""
+#            member = discord.Member()
+#            for mem in client.get_all_members():
+#                if mem.status() != "offline":
+#                    count += 1
+#                    mem = str(mem)
+#                    if count != 1:
+#                        members += ', '
+#                    members += mem
+#            final = "**Total : {count1} online members.**".format(count1=count)
+#            client.send_message(message.channel, memberson)
+#            client.send_message(message.channel, final)
+#            print(str("{} by {}".format(message.content,message.author)))
+#            logdis(message)
         # Uptime
         elif messagestr.startswith("!uptime"):
             deletemsg(message)
             time_elapsed = time.time() - start_time
             converted = gettime(time_elapsed)
             client.send_message(message.channel, "<@" + disauthor.id + "> **Uptime: {upt} **".format(upt=converted))
-            print(str("{msg} by {usr}".format(msg=message.content,usr=message.author)))
+            print(str("{} by {}".format(message.content,message.author)))
             logdis(message)
-        # Gets you an invite / todo fix this
-        elif messagestr.startswith("!getinvite"):
-            disinvite = client.create_invite(message.server)
-            deletemsg(message)
-            client.send_message(message.channel,"<@{usr}>, here is your code : {code} that expires in 24 hrs.".format(usr=disauthor.id,code=str(disinvite)))
-            print(str("{msg} by {usr}".format(msg=message.content,usr=message.author)))
-            logdis(message)
+        # Says something
+        elif messagestr.startswith("ayybot say"):
+            msgcut = messagestr
         # Returns mentioned user's avatar, or if no mention present, yours.
         elif messagestr.startswith("!avatar"):
             deletemsg(message)
@@ -215,7 +224,7 @@ def on_message(message):
             else:
                 user2 = message.author
                 client.send_message(message.channel,"<@" + disauthor.id + ">'s avatar: " + user2.avatar_url())
-            print(str("{msg} by {usr}".format(msg=message.content,usr=message.author)))
+            print(str("{} by {}".format(message.content,message.author)))
             logdis(message)
         # Decides between two or more words
         elif messagestr.startswith("!decide"):
@@ -226,7 +235,7 @@ def on_message(message):
             else:
                 client.send_message(message.channel, "<@" + disauthor.id + "> I have decided: " + msg3[1])
                 logdis(message)
-            print(str("{msg} by {usr}".format(msg=message.content,usr=message.author)))
+            print(str("{} by {}".format(message.content,message.author)))
         # Answers your question - 8ball style
         elif messagestr.startswith("!8ball"):
             deletemsg(message)
@@ -234,15 +243,14 @@ def on_message(message):
             answer = eightball[randint(1,len(eightball))]
             client.send_message(message.channel,"<@" + disauthor.id + "> asked : " + msg6 + "\n**" + answer + "**")
             logdis(message)
-            print(str("{msg} by {usr}".format(msg=message.content,usr=message.author)))
+            print(str("{} by {}".format(message.content,message.author)))
         # Returns a gif matching the name from Giphy (!gif <name>)
         elif messagestr.startswith("!gif"):
-            deletemsg(message)
             msg2 = str(message.content.lower()[4:])
             img = str(translate(msg2))
             client.send_message(message.channel, "<@" + disauthor.id + "> " + img)
             logdis(message)
-            print(str("{msg} by {usr}".format(msg=message.content,usr=message.author)))
+            print(str("{} by {}".format(message.content,message.author)))
         # Returns user info (name,id,discriminator and avatar url)
         elif messagestr.startswith("!user"):
             deletemsg(message)
@@ -257,14 +265,14 @@ def on_message(message):
             else:
                 client.send_message(message.channel, "Please mention someone to display their info")
             logdis(message)
-            print(str("{msg} by {usr}".format(msg=message.content,usr=message.author)))
+            print(str("{} by {}".format(message.content,message.author)))
         # Returns a random quote
         elif messagestr.startswith("!quote"):
             deletemsg(message)
             answer = quotes[randint(1,len(quotes))]
             client.send_message(message.channel,answer)
             logdis(message)
-            print(str("{msg} by {usr}".format(msg=message.content,usr=message.author)))
+            print(str("{} by {}".format(message.content,message.author)))
         # For managing roles
         elif messagestr.startswith("!role"):
             role1 = discord.utils.find(lambda role: role.name == 'serveradmins', message.channel.server.roles)
@@ -296,6 +304,17 @@ def on_message(message):
             else:
                 logdis(message)
                 print(str("{msg} by {usr}, but incorrect permissions".format(msg=message.content,usr=message.author)))
+        elif messagestr.startswith("!wiki"):
+            try:
+                wikipage = wikipedia.summary(str(messagestr[5:]),sentences=2)
+                client.send_message(message.channel,"*Wikipedia definition for* **" + messagestr[6:] + "** *:*\n" + wikipage)
+            except wikipedia.exceptions.PageError:
+                client.send_message(message.channel,"No definitions for " + messagestr[6:] + " were found")
+        elif messagestr.startswith("!urban"):
+            query = messagestr[6:]
+            define = requests.get("http://www.urbandictionary.com/define.php?term={}".format(query))
+            purehtml = BeautifulSoup(define.content, "html.parser")
+            client.send_message(message.channel,"*Urban definition for* **" + messagestr[6:] + "** *:*" + purehtml.find("div",attrs={"class":"meaning"}).text)
     except discord.InvalidArgument:
         print("Error -3 : InvalidArgument")
         client.send_message(message.channel, "Error -3 : InvalidArgument")
@@ -327,8 +346,8 @@ def on_member_join(member):
 
 # TIME
 start_time = time.time()
-# Write logs
-if parser.getboolean("SettingsOne","WriteLogs") == 1:
+# Write log line ----
+if parser.getboolean("Settings","WriteLogs") == 1:
     with open('log.txt','a') as file:
         file.write("\n------------------------------\n".format(vs=botvs))
